secured website after hack, password protection, csrf protection, tag/script injection protection

+++ health/bms/delete.php
@@ -1,6 +1,11 @@
 <?php
 include "../../inc/includes.php";
 
+if (!isset($_SESSION['user'])) {
+    header("Location: /login.php");
+    exit;
+}
+
 $id = isset($_REQUEST['id']) ? intval($_REQUEST['id']) : 0;
 
 $sql = "DELETE FROM vnd_bills_charge_categories WHERE id = :id ";
